ml_engine/requirements.txt
pandas==2.2.0
numpy==1.26.3
xgboost==2.0.3
scikit-learn==1.4.0
supabase>=2.4.0
python-dotenv==1.0.1
skl2onnx==1.16.0
onnx==1.15.0
onnxruntime==1.17.0
onnxmltools==1.12.0
